extract HA setup to a reusable utility and dynamically reduce resources on constrained clusters
